+++ app/common/locales/en.json
@@ -98,7 +98,7 @@
         "who_can_see_this" : "Who can see this?",
         "add" : "Add",
         "add_new_category" : "+ Add new category",
-        "sort_by": "Sort By",
+        "sort_by": "Sort By ...",
         "newest_first": "Newest first",
         "oldest_first": "Oldest first"
     },
