+++ app/main/posts/posts-module.js
@@ -64,6 +64,7 @@ angular.module('ushahidi.posts', [])
 .directive('postActiveFilters', require('./views/filters/active-filters.directive.js'))
 .directive('postActiveSearchFilters', require('./views/filters/active-search-filters.directive.js'))
 .service('PostFilters', require('./views/post-filters.service.js'))
+.service('FilterTransformers', require('./views/filters/filter-transformers.service.js'))
 .directive('filtersDropdown', require('./views/filters/filters-dropdown.directive.js'))
 .directive('filterSavedSearch', require('./views/filters/filter-saved-search.directive.js'))
 .directive('sortAndFilterCounter', require('./views/filters/sort-and-filter-counter.directive.js'))
