+++ app/main/posts/views/filters/filter-date.html
@@ -1,5 +1,5 @@
 <fieldset>
-    <legend translate>global_filter.date_range</legend>
+    <label translate>global_filter.date_range</label>
     <div class="form-field date">
         <label class="hidden" translate="global_filter.filter_tabs.created_after">Start date</label>
         <div class="input-with-icon">
