+++ app/main/posts/views/filters/filter-posts.html
@@ -1,11 +1,11 @@
-<div>
+
 <form name="postFiltersFormOpen"
       dropdown
       role="search"
       class="searchbar"
       auto-close="outsideClick"
       ng-submit="applyFilters()"
-      ng-model-options="{ updateOn : 'submit' }"
+      ng-model-options="{ updateOn : 'submit radio' }"
       data-message="search"
 >
     <filter-searchbar model="filters.q"></filter-searchbar>
@@ -15,14 +15,6 @@
             <!-- DEVNOTE translate -->
             <sort-and-filter-counter></sort-and-filter-counter>
         </a>
-        <!-- <a id="togglePostSearchFiltersOpen" class="button searchbar-options-filter" ng-click="filtersDropdownToggle = !filtersDropdownToggle">
-            <span class="button-label" translate>app.filters</span>
-            <svg class="iconic">
-                <use xlink:href="/img/iconic-sprite.svg#chevron-bottom">
-                </use>
-            </svg>
-        </a> -->
     </div>
-</form>
-<filters-dropdown filters-var="filters" ng-model="filters" cancel="cancel" filters-dropdown-toggle="filtersDropdownToggle" apply-filters="applyFilters"></filters-dropdown>
-</div>
+    <filters-dropdown filters-var="filters" ng-model="filters" cancel="cancel" filters-dropdown-toggle="filtersDropdownToggle" apply-filters="applyFilters"></filters-dropdown>
+</form><!-- this is the guilty ng-model-options -->
